requirements.txt
discord.py
asyncio
bs4
lxml
requests
